minimize the test and a sketchy fix
