+++ framework/src/main/java/org/checkerframework/common/wholeprograminference/WholeProgramInferenceImplementation.java
@@ -804,7 +804,17 @@ public class WholeProgramInferenceImplementation<T> implements WholeProgramInfer
 
     AnnotatedTypeMirror atmFromStorage =
         storage.atmFromStorageLocation(rhsATM.getUnderlyingType(), annotationsToUpdate);
-    updateAtmWithLub(rhsATM, atmFromStorage);
+    // It is possible that rhsATM is a type variable but atmFromStorage is not, which
+    // would cause a ClassCastException in updateAtmWithLub(). This can
+    // happen for example when a local variable of type ? extends T is assigned to a
+    // field of type T, such as in the test ainfer-index/non-annotated/Dataset6Crash.java.
+    try {
+      updateAtmWithLub(rhsATM, atmFromStorage);
+    } catch (ClassCastException c) {
+      AnnotatedTypeMirror compatibleRHSAtm =
+          AnnotatedTypes.asSuper(this.atypeFactory, rhsATM, atmFromStorage);
+      updateAtmWithLub(compatibleRHSAtm, atmFromStorage);
+    }
     if (lhsATM instanceof AnnotatedTypeVariable) {
       Set<AnnotationMirror> upperAnnos =
           ((AnnotatedTypeVariable) lhsATM).getUpperBound().getEffectiveAnnotations();
